// src/types/SearchMovies.ts
export type SearchMovieType = {
    adult: boolean
    backdrop_path: string
    genre_ids: number[]
    id: number
    original_language: string
    original_title: string
    overview: string
    popularity: number
    poster_path: string | null
    release_date: string
    title: string
    video: boolean
    vote_average: number
    vote_count: number
  }
  
  export type SearchMovieResponseType = {
    results: SearchMovieType[]
  } & SearchMoviePaginationType
  
  export type SearchMoviePaginationType = {
    page: number
    total_pages: number
    total_results: number
  }